Refactor helper outside
