+++ src/main/galmart/extractor/BuyFromMarketExtractor.java
@@ -5,9 +5,8 @@ import java.util.List;
 import com.fs.starfarer.api.campaign.econ.CommodityOnMarketAPI;
 import com.fs.starfarer.api.campaign.econ.EconomyAPI;
 import com.fs.starfarer.api.campaign.econ.MarketAPI;
-import com.fs.starfarer.api.impl.campaign.submarkets.OpenMarketPlugin;
 
-public class BuyFromMarketExtractor extends SortableMarketExtractor {
+public class BuyFromMarketExtractor extends MarketExtractor {
 
     public BuyFromMarketExtractor(String commodityId, EconomyAPI economy) {
         super(commodityId, economy);
@@ -19,6 +18,12 @@ public class BuyFromMarketExtractor extends SortableMarketExtractor {
     }
 
     @Override
+    public List<MarketAPI> getMarkets() {
+        sortMarkets();
+        return markets;
+    }
+
+    @Override
     public List<Object[]> getRows() {
         sortMarkets();
         return super.getRows();
@@ -26,24 +31,18 @@ public class BuyFromMarketExtractor extends SortableMarketExtractor {
 
     @Override
     public float getPrice(MarketAPI market) {
-        float econUnit = commoditySpec.getEconUnit();
-        return market.getSupplyPrice(commodityId, econUnit, true) / econUnit;
+        return helper.getSupplyPrice(market);
     }
 
     @Override
     protected Object[] getRow(MarketAPI market) {
         CommodityOnMarketAPI commodity = market.getCommodityData(commodityId);
-        int available = getAvailable(commodity);
+        int available = helper.getAvailable(commodity);
+        // TODO replace with filter
         if (available <= 0) {
             return null;
         }
         int excess = commodity.getExcessQuantity();
         return getRow(market, commodity, available, excess);
-    }
-
-    private int getAvailable(CommodityOnMarketAPI commodity) {
-        int available = OpenMarketPlugin.getApproximateStockpileLimit(commodity);
-        available += commodity.getPlayerTradeNetQuantity();
-        return available;
     }
 }
