+++ src/main/galmart/extractor/ExtractorHelper.java
@@ -1,0 +1,94 @@
+package galmart.extractor;
+
+import java.awt.Color;
+
+import com.fs.starfarer.api.campaign.FactionAPI;
+import com.fs.starfarer.api.campaign.StarSystemAPI;
+import com.fs.starfarer.api.campaign.econ.CommodityOnMarketAPI;
+import com.fs.starfarer.api.campaign.econ.CommoditySpecAPI;
+import com.fs.starfarer.api.campaign.econ.MarketAPI;
+import com.fs.starfarer.api.impl.campaign.submarkets.OpenMarketPlugin;
+import com.fs.starfarer.api.util.Misc;
+
+public class ExtractorHelper {
+
+    private String commodityId;
+    private CommoditySpecAPI commoditySpec;
+
+    public ExtractorHelper(String commodityId, CommoditySpecAPI commoditySpec) {
+        this.commodityId = commodityId;
+        this.commoditySpec = commoditySpec;
+    }
+
+    public int getAvailable(CommodityOnMarketAPI commodity) {
+        int available = OpenMarketPlugin.getApproximateStockpileLimit(commodity);
+        available += commodity.getPlayerTradeNetQuantity();
+        return available;
+    }
+
+    public Color getClaimingFactionColor(MarketAPI market) {
+        FactionAPI faction = Misc.getClaimingFaction(market.getPrimaryEntity());
+        if (faction == null) {
+            return Misc.getGrayColor();
+        }
+        return faction.getColor();
+    }
+
+    public int getDemand(MarketAPI market, CommodityOnMarketAPI commodity) {
+        int demandIcons = commodity.getMaxDemand();
+        if (!commodity.getCommodity().isPrimary()) {
+            CommodityOnMarketAPI primary = market.getCommodityData(commodity.getCommodity().getDemandClass());
+            demandIcons = primary.getMaxDemand();
+        }
+        int demand = (int) (commodity.getCommodity().getEconUnit() * demandIcons);
+        demand -= commodity.getPlayerTradeNetQuantity();
+        return demand;
+    }
+
+    public float getDemandPrice(MarketAPI market) {
+        float econUnit = commoditySpec.getEconUnit();
+        return market.getDemandPrice(commodityId, econUnit, true) / econUnit;
+    }
+
+    public Color getExcessColor(int excess) {
+        if (excess > 0) {
+            return Misc.getPositiveHighlightColor();
+        }
+        if (excess < 0) {
+            return Misc.getNegativeHighlightColor();
+        }
+        return Misc.getGrayColor();
+    }
+
+    public String getExcessValue(int excess) {
+        if (excess > 0) {
+            return Misc.getWithDGS(excess);
+        }
+        if (excess < 0) {
+            return Misc.getWithDGS(-excess);
+        }
+        return "---";
+    }
+
+    public String getDistance(MarketAPI market) {
+        float distance = Misc.getDistanceToPlayerLY(market.getPrimaryEntity());
+        return String.format("%.1f", distance);
+    }
+
+    public String getLocation(MarketAPI market) {
+        return market.getName() + " - " + market.getFaction().getDisplayName();
+    }
+
+    public float getSupplyPrice(MarketAPI market) {
+        float econUnit = commoditySpec.getEconUnit();
+        return market.getSupplyPrice(commodityId, econUnit, true) / econUnit;
+    }
+
+    public String getSystemName(MarketAPI market) {
+        StarSystemAPI starSystem = market.getStarSystem();
+        if (starSystem == null) {
+            return "Hyperspace";
+        }
+        return starSystem.getBaseName();
+    }
+}
