+++ src/main/galmart/extractor/MarketFactory.java
@@ -1,0 +1,8 @@
+package galmart.extractor;
+
+public class MarketFactory {
+
+    // add market generation
+    // sorting
+    // filtering
+}
