+++ src/main/galmart/filter/Filter.java
@@ -1,0 +1,6 @@
+package galmart.filter;
+
+public interface Filter<T> {
+
+    public boolean accept(T object);
+}
